Move validation to schema file.
