+++ src/constants/__tests__/configSchema.spec.js
@@ -1,0 +1,76 @@
+import { validateConfig } from '../configSchema';
+
+describe('config', () => {
+  describe('validateConfig', () => {
+    it('should not throw if no errors', () => {
+      const config = {
+        foo: 'bar',
+        backend: { name: 'bar' },
+        media_folder: 'baz',
+        collections: [{
+          name: 'posts',
+          label: 'Posts',
+          folder: '_posts',
+          fields: [{ name: 'title', label: 'title', widget: 'string' }],
+        }],
+      };
+      expect(() => {
+        validateConfig(config);
+      }).not.toThrowError();
+    });
+
+    it('should throw if backend is not defined in config', () => {
+      expect(() => {
+        validateConfig({ foo: 'bar' });
+      }).toThrowError("config should have required property 'backend'");
+    });
+
+    it('should throw if backend name is not defined in config', () => {
+      expect(() => {
+        validateConfig({ foo: 'bar', backend: {} });
+      }).toThrowError("'backend' should have required property 'name'");
+    });
+
+    it('should throw if backend name is not a string in config', () => {
+      expect(() => {
+        validateConfig({ foo: 'bar', backend: { name: { } } });
+      }).toThrowError("'backend.name' should be string");
+    });
+
+    it('should throw if media_folder is not defined in config', () => {
+      expect(() => {
+        validateConfig({ foo: 'bar', backend: { name: 'bar' } });
+      }).toThrowError("config should have required property 'media_folder'");
+    });
+
+    it('should throw if media_folder is not a string in config', () => {
+      expect(() => {
+        validateConfig({ foo: 'bar', backend: { name: 'bar' }, media_folder: {} });
+      }).toThrowError("'media_folder' should be string");
+    });
+
+    it('should throw if collections is not defined in config', () => {
+      expect(() => {
+        validateConfig({ foo: 'bar', backend: { name: 'bar' }, media_folder: 'baz' });
+      }).toThrowError("config should have required property 'collections'");
+    });
+
+    it('should throw if collections not an array in config', () => {
+      expect(() => {
+        validateConfig({ foo: 'bar', backend: { name: 'bar' }, media_folder: 'baz', collections: {} });
+      }).toThrowError("'collections' should be array");
+    });
+
+    it('should throw if collections is an empty array in config', () => {
+      expect(() => {
+        validateConfig({ foo: 'bar', backend: { name: 'bar' }, media_folder: 'baz', collections: [] });
+      }).toThrowError("'collections' should NOT have less than 1 items");
+    });
+
+    it('should throw if collections is an array with a single null element in config', () => {
+      expect(() => {
+        validateConfig({ foo: 'bar', backend: { name: 'bar' }, media_folder: 'baz', collections: [null] });
+      }).toThrowError("'collections[0]' should be object");
+    });
+  });
+});
